Moved motor and inout state definitions to separate file so that it will be easier to use from other modules.
